+++ src/Main.hs
@@ -1,3 +1,7 @@
+import           Chat.Bot.Answer.Guess
+import           Chat.Bot.Answer.Cipher
+import           Chat.Bot.Answer.Calculator
+import           Chat.Bot.Answer.TicTacToe
 import           Chat.Bot.Cipher
 import           Chat.Bot.Calculator
 import           Chat.Bot.Guess
